+++ NEWS.md
@@ -1,3 +1,12 @@
+# slideutils 0.1.4
+
+* added fireworks animated GIF meme
+* `su_meme()`: `height` argument can now be 'default' (no resize, recommended for animated GIFs)
+
+# slideutils 0.1.3
+
+* fixed a typo in gfs-report.Rmd
+
 # slideutils 0.1.2
 
 * `su_meme()`: fixed a typo in the code that would cause an error